Module Jude__Beh

Beh module have higher level interfaces over some primitives from Arbiter.

module Supervisor : functor (_ : Jude.Arbiter.ARBITER) -> sig ... end

Module used for supervising actor. It uses monitor primitives.

module Resolver : functor (_ : Jude.Arbiter.ARBITER) -> sig ... end

Resolver is used for doing actor pid lookups based on name.