Module Jude__Beh.Supervisor

Module used for supervising actor. It uses monitor primitives.

Parameters

_ : Jude.Arbiter.ARBITER

Signature

type recipe = {

name : string;

public : bool;

beh : Jude.Actor.beh;

}

Recipe which is used for creating

type policy =

| One_for_one

| All_for_one
One_for_one is used when one to one actor should be replaced. When All_for_one set all actors supervised will be restarted.

type t

val create : ?⁠policy:policy -> recipe list -> t

val run : t -> Jude.Actor.beh

Run is a behaviour function which should be used for spawning supervisors.